% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beatr_function.R
\name{beatr}
\alias{beatr}
\title{A nastyverse function}
\usage{
beatr(sound = "ecoroms")
}
\arguments{
\item{sound}{character string specifiying which built in sound to be played. Defaults to ecoroms.}
}
\description{
This function allows you play some nasty beats to accompany your code.
}
\examples{
# Play a nasty sound
beatr("tubular")
# Update all packages and hear Good Morning Vietnam! when it is ready
beatr("ecoroms")
}
\keyword{cats}
