% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beatr_function.R
\name{beatr}
\alias{beatr}
\title{A nastyverse function}
\usage{
beatr(sound)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
# Play a nasty sound beatr("tubular")
}
\keyword{cats}
